requirements.txt
# Install with: 'pip install -r requirements.txt'

Django==1.9.2
djangorestframework==3.3.3
MySQL-python==1.2.5
django-filter
django-ipware
pillow
